# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: DBackup API
  version: 2.0.0
  description: |
    REST API for DBackup - a self-hosted database backup automation platform with encryption, compression, and smart retention.

    ## Authentication

    DBackup supports two authentication methods:

    - **Session Authentication** - Used automatically when logged in via the web UI. Session cookies are sent with each request.
    - **API Key Authentication** - For scripts, CI/CD pipelines, and external integrations. Create an API key under **Access Management → API Keys**.

    > API keys do not inherit SuperAdmin privileges. Only explicitly assigned permissions are available.

    ## Rate Limits

    | Request Type | Default Limit |
    |---|---|
    | Authentication (login, etc.) | 5/min per IP |
    | GET requests | 100/min per IP |
    | POST / PUT / DELETE | 20/min per IP |

    Rate limits are configurable in Settings → Rate Limits.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: "{protocol}://{host}/api"
    description: DBackup Instance
    variables:
      protocol:
        default: https
        enum: [http, https]
      host:
        default: your-dbackup-instance.com
security:
  - BearerAuth: []
  - SessionAuth: []
...
paths:
...
  /jobs/{id}/run:
    post:
      tags: [Jobs]
      summary: Trigger Job
      description: |
        Manually trigger a backup job execution. The job is added to the execution queue.
        If the max concurrent jobs limit is reached, it will remain in `Pending` status until a slot is available.

        Use the returned `executionId` to poll execution status via the Executions endpoint.

        **Required permission:** `jobs:execute`
      operationId: triggerJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Job queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  executionId:
                    type: string
                    example: "clx1abc..."
                  message:
                    type: string
                    example: "Job queued successfully"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /executions/{id}:
    get:
      tags: [Executions]
      summary: Get Execution
      description: |
        Poll the status of a running or completed execution.
        Use `includeLogs=true` to include full execution log entries.

        **Required permission:** `history:read`
      operationId: getExecution
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: includeLogs
          in: query
          schema:
            type: boolean
            default: false
          description: Include full execution log entries
      responses:
        "200":
          description: Execution details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/ExecutionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /storage/{id}/download:
    get:
      tags: [Storage Explorer]
      summary: Download Backup
      description: Downloads a backup file. Optionally decrypts the file if it was encrypted.
      operationId: downloadBackup
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: file
          in: query
          required: true
          schema:
            type: string
          description: Backup file path
        - name: decrypt
          in: query
          schema:
            type: boolean
            default: true
          description: Auto-decrypt the file
      responses:
        "200":
          description: Binary file stream
          headers:
            Content-Disposition:
              schema:
                type: string
                example: "attachment; filename=\"mydb_2025-01-15.sql.gz\""
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /storage/{id}/download-url:
    post:
      tags: [Storage Explorer]
      summary: Generate Download URL
      description: |
        Creates a temporary, single-use download link that works without authentication.
        The token expires after 5 minutes and is consumed on first use.
      operationId: generateDownloadUrl
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  description: Backup file path
                  example: "backups/mydb_2025-01-15.sql.gz.enc"
                decrypt:
                  type: boolean
                  default: true
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Download URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  url:
                    type: string
                    example: "https://your-instance.com/api/storage/public-download?token=..."
                  expiresIn:
                    type: string
                    example: "5 minutes"
                  singleUse:
                    type: boolean
                    example: true
...
  /storage/{id}/analyze:
    post:
      tags: [Storage Explorer]
      summary: Analyze Backup
      description: Extract database names from a backup file (used for restore UI).
      operationId: analyzeBackup
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  description: Backup file path
                type:
                  type: string
                  description: Database type hint (e.g., `mysql`)
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Databases in backup
          content:
            application/json:
              schema:
                type: object
                properties:
                  databases:
                    type: array
                    items:
                      type: string
                    example: ["mydb", "app_production"]
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API Key authentication. Create an API key under **Access Management → API Keys**.

        Format: `dbackup_xxxxxxxx`
    SessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie (browser login)
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  responses:
    Unauthorized:
      description: Missing, invalid, disabled, or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Valid credentials but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Human-readable error message"
...
    ExecutionDetail:
      type: object
      properties:
        id:
          type: string
        jobId:
          type: string
        jobName:
          type: string
        type:
          type: string
          enum: [Backup, Restore]
        status:
          $ref: "#/components/schemas/ExecutionStatus"
        progress:
          type: integer
          minimum: 0
          maximum: 100
        stage:
          type: string
          description: Current pipeline stage
          example: "Uploading"
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          # nullable handled via type array
        duration:
          type: [number, "null"]
          description: Duration in seconds
        size:
          type: [integer, "null"]
          description: Backup size in bytes
        path:
          type: [string, "null"]
        error:
          type: [string, "null"]
        logs:
          type: array
          items:
            type: object
          description: Log entries (only if `includeLogs=true`)
    ExecutionStatus:
      type: string
      enum: [Pending, Running, Success, Failed, Partial, Cancelled]
      description: |
        - `Pending` - Queued, waiting for an execution slot
        - `Running` - Currently executing
        - `Success` - Completed successfully
        - `Failed` - Execution failed
        - `Partial` - Some destinations succeeded, others failed
        - `Cancelled` - Cancelled by a user before or during execution

    # ─── Adapters ────────────────────────────────────────────────
...